use device as en enum and default to mobile
