+++ TopsortAnalytics/src/main/java/com/topsort/analytics/model/auctions/AuctionResponse.kt
@@ -70,19 +70,3 @@ data class AuctionResponse private constructor(
     }
 }
 
-enum class EntityType {
-    PRODUCT,
-    VENDOR,
-    BRAND,
-    URL;
-
-    companion object {
-        fun fromValue(value: String): EntityType = when (value) {
-            "product" -> PRODUCT
-            "vendor" -> VENDOR
-            "brand" -> BRAND
-            "url" -> URL
-            else -> throw IllegalArgumentException("not valid entity type: $value")
-        }
-    }
-}
