+++ TopsortAnalytics/src/main/java/com/topsort/analytics/model/auctions/Device.kt
@@ -1,0 +1,6 @@
+package com.topsort.analytics.model.auctions
+
+enum class Device {
+    DESKTOP,
+    MOBILE
+}
